+++ release-notes.md
@@ -8,6 +8,9 @@
 
 #### 🎉 Major rework to make "Merge & Export" less destructive (#29, #36)
 
+- Fixed a Merge & Export bug. Forgot to run the "apply modifiers on object with shapekeys" during export
+  - _(using [przemir/ApplyModifierForObjectWithShapeKeys](https://github.com/przemir/ApplyModifierForObjectWithShapeKeys))_
+
 ❌ _In previous versions,_ I warned you **_NOT TO SAVE_** after pressing that dangerous "Merge & Export" button.
 
 ✅ _Now,_ you can safely press "Merge & Export", and then continue working. It performs the merge in a temporary Blender scene instead of your main scene.